==> interpolation_notebooks/2_Data_Imputation.ipynb <==
# Necessary libraries

import warnings
warnings.filterwarnings('ignore')
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy.interpolate import lagrange
from scipy.interpolate import CubicSpline, InterpolatedUnivariateSpline, Akima1DInterpolator
from datetime import datetime, timedelta
from statsmodels.tsa.api import ExponentialSmoothing, SimpleExpSmoothing, Holt
import statsmodels.api as sm
#from datetime import timedelta
#import datetime
#import fbprophet
#from matplotlib import pyplot
from pandas import to_datetime
from statsmodels.tsa.arima_model import ARIMA
#from fbprophet import Prophet
#from pmdarima.arima import auto_arima
#import pmdarima
from scipy import interpolate
#from fbprophet import Prophet
from matplotlib import pyplot
from tensorflow.keras.preprocessing.sequence import TimeseriesGenerator
from tensorflow.keras import Sequential, layers, callbacks
from tensorflow.keras.layers import Dense, LSTM, Dropout, GRU, Bidirectional
import plotly.graph_objs as go
import sklearn.model_selection
from scipy import stats
from keras.callbacks import EarlyStopping
from sklearn.preprocessing import MinMaxScaler, StandardScaler
from tensorflow import keras
from imblearn.over_sampling import SMOTE
from scipy.interpolate import CubicSpline as CS
from sklearn.ensemble import RandomForestClassifier
from scipy import stats
from sklearn.model_selection import GridSearchCV,\
RandomizedSearchCV

#!pip install pystan==2.19.1.1 convertdate==2.4.0 lunarcalendar==0.0.9 holidays==0.13 fbprophet
#!pip install plotly

Wound_assessment = pd.read_csv("2_data_for_imputation_may11.csv")
Wound_assessment=Wound_assessment.drop(["Unnamed: 0"],axis=1)
Wound_assessment

Wound_assessment["PatientID"]=Wound_assessment["PatientID"].astype("int")
#Wound_assessment["SmokingStatus"]=Wound_assessment["SmokingStatus"].astype("int")
Wound_assessment

df3=Wound_assessment
df3["Area"]=df3["Length"]*df3["Width"]
df3=df3.reset_index(drop=True)
Wound_assessment=df3
Wound_assessment

Wound_assessment["SmokingStatus"]=Wound_assessment["SmokingStatus"].replace({np.nan:0.0})
Wound_assessment["BMI"]=Wound_assessment["BMI"].replace({np.nan:Wound_assessment["BMI"].mean()})
Wound_assessment

list_patients=Wound_assessment["PatientID"].unique()
print("There are",len(list_patients),"number of patients whose wound assessment information is found in the merged dataframe.")
list_count_patient_for_forecasting=[]
unique_wound_number=0
for i in range(len(list_patients)):
    #print(i)
    patient=list_patients[i]
    df_patient=Wound_assessment.query("PatientID == @patient")
    if df_patient.shape[0]>1:
        #print(i)
        wound_list_patient=df_patient["WoundNumber"].unique().tolist()
        print("For patient",i,"number of unique wounds",len(wound_list_patient))
        unique_wound_number=unique_wound_number+len(wound_list_patient)
        list_count_patient_for_forecasting.append(i)
        #display(df_patient)
print("Forecasting can be done for",len(list_count_patient_for_forecasting),"number of patients, because they have more than one wound assessments. Others do not.")
print("Number of unique wounds for",len(list_count_patient_for_forecasting),"patients is",unique_wound_number)

## Making two separate dataframes for numeric and categorical variables and imputing the values

Wound_assessment_numeric=Wound_assessment[["PatientID", "DateAdded_New","WoundNumber","Length","Width","Age","BMI"]]
Wound_assessment_numeric= Wound_assessment_numeric.groupby('PatientID').agg('mean')
Wound_assessment_numeric=Wound_assessment_numeric.reset_index()
Wound_assessment_numeric

Wound_assessment_categorical=Wound_assessment[["PatientID","PatientGender","RaceCodeList","patientEthnicity","SmokingStatus"]]
Wound_assessment_categorical= Wound_assessment_categorical.groupby('PatientID').agg('max')
Wound_assessment_categorical=Wound_assessment_categorical.reset_index()
Wound_assessment_categorical

### Merging and Replacing all Unknowns with np.nan

#pd.merge(data_numeric, data_coded,left_index=True, right_index=True)
df2 = pd.merge(Wound_assessment_numeric, Wound_assessment_categorical, on='PatientID')
df2=df2.reset_index()
df2=df2.drop(["index"],axis=1)
df2

print("Value counts for Gender:")
print(df2["PatientGender"].value_counts())
print("\n")
print("Value counts for Race:")
print(df2["RaceCodeList"].value_counts())
print("\n")
print("Value counts for Ethnicity:")
print(df2["patientEthnicity"].value_counts())

df2=df2.replace({"Unknown":np.nan})
df2["PatientGender"]=df2["PatientGender"].replace({"M":0,"F":1})
df2["RaceCodeList"]=df2["RaceCodeList"].replace({"White":0,"Black":1,"Asian":2,"ASKU":3,
                                                 "American Indian or Alaska Native":4,
                                                 "Native Hawaiian or Other Pacific Islander":5,
                                                 "Other":6})
df2["patientEthnicity"]=df2["patientEthnicity"].replace({"Non-Hispanic":0,"Hispanic":1,"Patient Declined":2,"ASKU":3})

print("Number of PatientGender need be imputed:",df2["PatientGender"].shape[0]-df2["PatientGender"].dropna().shape[0])
print("Number of RaceCodeList need be imputed:",df2["RaceCodeList"].shape[0]-df2["RaceCodeList"].dropna().shape[0])
print("Number of patientEthnicity need be imputed:",df2["patientEthnicity"].shape[0]-df2["patientEthnicity"].dropna().shape[0])

## Imputing PatientGender

#default without hyperparameter tuning

#pd.merge(data_numeric, data_coded,left_index=True, right_index=True)
df2 = pd.merge(Wound_assessment_numeric, Wound_assessment_categorical, on='PatientID')
df2=df2.reset_index()
df2=df2.drop(["index"],axis=1)
df2

print("Value counts for Gender:")
print(df2["PatientGender"].value_counts())
print("\n")
print("Value counts for Race:")
print(df2["RaceCodeList"].value_counts())
print("\n")
print("Value counts for Ethnicity:")
print(df2["patientEthnicity"].value_counts())

df2=df2.replace({"Unknown":np.nan})
df2["PatientGender"]=df2["PatientGender"].replace({"M":0,"F":1})
df2["RaceCodeList"]=df2["RaceCodeList"].replace({"White":0,"Black":1,"Asian":2,"ASKU":3,
                                                 "American Indian or Alaska Native":4,
                                                 "Native Hawaiian or Other Pacific Islander":5,
                                                 "Other":6})
df2["patientEthnicity"]=df2["patientEthnicity"].replace({"Non-Hispanic":0,"Hispanic":1,"Patient Declined":2,"ASKU":3})

print("Number of PatientGender need be imputed:",df2["PatientGender"].shape[0]-df2["PatientGender"].dropna().shape[0])
print("Number of RaceCodeList need be imputed:",df2["RaceCodeList"].shape[0]-df2["RaceCodeList"].dropna().shape[0])
print("Number of patientEthnicity need be imputed:",df2["patientEthnicity"].shape[0]-df2["patientEthnicity"].dropna().shape[0])
#data=df2
df3=df2
#df2=df2.drop(["index"],axis=1)
df3=df3.sort_values(by=['PatientGender'])

df3["Length"]=df3["Length"].fillna(value=df3["Length"].mean())
df3["Width"]=df3["Width"].fillna(value=df3["Width"].mean())
#df3["Depth"]=df3["Depth"].fillna(value=df3["Depth"].mean())
df3["RaceCodeList"]=df3["RaceCodeList"].fillna(value=df3["RaceCodeList"].mean())
df3["patientEthnicity"]=df3["patientEthnicity"].fillna(value=df3["patientEthnicity"].mean())



test_set=df3.tail(df3["PatientGender"].shape[0]-df3["PatientGender"].dropna().shape[0])
test_set
#new_df = df.drop(df1.index)
train_set=df3.drop(test_set.index)
train_set

X_train=train_set.drop(["PatientID","PatientGender"],axis=1)
X_test=test_set.drop(["PatientID","PatientGender"],axis=1)
y_train=train_set["PatientGender"]

X_train=X_train.values
y_train=y_train.values

# define oversampling strategy
SMOTE = SMOTE()

# fit and apply the transform
X_train_SMOTE, y_train_SMOTE = SMOTE.fit_resample(X_train, y_train)

rf = RandomForestClassifier()
rf.fit(X_train_SMOTE, y_train_SMOTE)

y_pred = rf.predict(X_test)
test_set["PatientGender"]=y_pred

new_df_gender=pd.concat([train_set,test_set],axis=0)
new_df_gender=new_df_gender[["PatientID","PatientGender"]]
new_df_gender.rename({'PatientGender': 'PatientGender_new'}, axis=1, inplace=True)


dff=pd.merge(df2,new_df_gender,on="PatientID")
dff=dff.drop(["PatientGender"],axis=1)
df2=dff
df2

print(train_set["PatientGender"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')
print(df2["PatientGender_new"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')

param_grid = {
    'n_estimators': [25, 50, 100, 150],
    'max_features': ['sqrt', 'log2', None],
    'max_depth': [3, 6, 9],
    'max_leaf_nodes': [3, 6, 9],
}

grid_search = GridSearchCV(RandomForestClassifier(),
                           param_grid=param_grid)
grid_search.fit(X_train_SMOTE, y_train_SMOTE)
print(grid_search.best_estimator_)

#with hyperparameter tuning
#data=df2
df3=df2
#df2=df2.drop(["index"],axis=1)
df3=df3.sort_values(by=['PatientGender'])

df3["Length"]=df3["Length"].fillna(value=df3["Length"].mean())
df3["Width"]=df3["Width"].fillna(value=df3["Width"].mean())
#df3["Depth"]=df3["Depth"].fillna(value=df3["Depth"].mean())
df3["RaceCodeList"]=df3["RaceCodeList"].fillna(value=df3["RaceCodeList"].mean())
df3["patientEthnicity"]=df3["patientEthnicity"].fillna(value=df3["patientEthnicity"].mean())



test_set=df3.tail(df3["PatientGender"].shape[0]-df3["PatientGender"].dropna().shape[0])
test_set
#new_df = df.drop(df1.index)
train_set=df3.drop(test_set.index)
train_set

X_train=train_set.drop(["PatientID","PatientGender"],axis=1)
X_test=test_set.drop(["PatientID","PatientGender"],axis=1)
y_train=train_set["PatientGender"]

X_train=X_train.values
y_train=y_train.values

# define oversampling strategy
SMOTE = SMOTE()

# fit and apply the transform
X_train_SMOTE, y_train_SMOTE = SMOTE.fit_resample(X_train, y_train)

model_grid = RandomForestClassifier(max_depth=6,
                                    max_features="log2",
                                    max_leaf_nodes=9,
                                    n_estimators=25)
model_grid.fit(X_train_SMOTE, y_train_SMOTE)

y_pred = rf.predict(X_test)
test_set["PatientGender"]=y_pred

new_df_gender=pd.concat([train_set,test_set],axis=0)
new_df_gender=new_df_gender[["PatientID","PatientGender"]]
new_df_gender.rename({'PatientGender': 'PatientGender_new'}, axis=1, inplace=True)


dff=pd.merge(df2,new_df_gender,on="PatientID")
dff=dff.drop(["PatientGender"],axis=1)
df2=dff
df2

print(train_set["PatientGender"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')
print(df2["PatientGender_new"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')

## Imputing RaceCodeList

from imblearn.over_sampling import SMOTE
#data=df2
df3=df2
#df2=df2.drop(["index"],axis=1)
df3=df3.sort_values(by=['RaceCodeList'])

df3["Length"]=df3["Length"].fillna(value=df3["Length"].mean())
df3["Width"]=df3["Width"].fillna(value=df3["Width"].mean())
#df3["Depth"]=df3["Depth"].fillna(value=df3["Depth"].mean())
df3["patientEthnicity"]=df3["patientEthnicity"].fillna(value=df3["patientEthnicity"].mean())



test_set=df3.tail(df3["RaceCodeList"].shape[0]-df3["RaceCodeList"].dropna().shape[0])
test_set
#new_df = df.drop(df1.index)
train_set=df3.drop(test_set.index)
train_set

X_train=train_set.drop(["PatientID","RaceCodeList"],axis=1)
X_test=test_set.drop(["PatientID","RaceCodeList"],axis=1)
y_train=train_set["RaceCodeList"]

X_train=X_train.values
y_train=y_train.values

# define oversampling strategy

SMOTE = SMOTE(k_neighbors=1)
#SMOTE=SMOTE()
# fit and apply the transform
#X_train_SMOTE, y_train_SMOTE = SMOTE.fit_resample(X_train, y_train)

X_train_SMOTE, y_train_SMOTE = SMOTE.fit_resample(X_train, y_train)


rf = RandomForestClassifier()
rf.fit(X_train_SMOTE, y_train_SMOTE)

y_pred = rf.predict(X_test)
test_set["RaceCodeList"]=y_pred

new_df_gender=pd.concat([train_set,test_set],axis=0)
new_df_gender=new_df_gender[["PatientID","RaceCodeList"]]
new_df_gender.rename({'RaceCodeList': 'RaceCodeList_new'}, axis=1, inplace=True)


dff=pd.merge(df2,new_df_gender,on="PatientID")
dff=dff.drop(["RaceCodeList"],axis=1)
df2=dff
df2

print(train_set["RaceCodeList"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')
print(df2["RaceCodeList_new"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')

## Imputing patientEthnicity

from imblearn.over_sampling import SMOTE
#data=df2
df3=df2
#df2=df2.drop(["index"],axis=1)
df3=df3.sort_values(by=['patientEthnicity'])

df3["Length"]=df3["Length"].fillna(value=df3["Length"].mean())
df3["Width"]=df3["Width"].fillna(value=df3["Width"].mean())
#df3["Depth"]=df3["Depth"].fillna(value=df3["Depth"].mean())



test_set=df3.tail(df3["patientEthnicity"].shape[0]-df3["patientEthnicity"].dropna().shape[0])
test_set
#new_df = df.drop(df1.index)
train_set=df3.drop(test_set.index)
train_set

X_train=train_set.drop(["PatientID","patientEthnicity"],axis=1)
X_test=test_set.drop(["PatientID","patientEthnicity"],axis=1)
y_train=train_set["patientEthnicity"]

X_train=X_train.values
y_train=y_train.values

# # define oversampling strategy
# SMOTE = SMOTE(k_neighbors=2)

# # fit and apply the transform
# X_train_SMOTE, y_train_SMOTE = SMOTE.fit_resample(X_train, y_train)

rf = RandomForestClassifier()
rf.fit(X_train, y_train)

y_pred = rf.predict(X_test)
test_set["patientEthnicity"]=y_pred

new_df_gender=pd.concat([train_set,test_set],axis=0)
new_df_gender=new_df_gender[["PatientID","patientEthnicity"]]
new_df_gender.rename({'patientEthnicity': 'patientEthnicity_new'}, axis=1, inplace=True)


dff=pd.merge(df2,new_df_gender,on="PatientID")
dff=dff.drop(["patientEthnicity"],axis=1)
df2=dff
df2

print(train_set["patientEthnicity"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')
print(df2["patientEthnicity_new"].value_counts(normalize=True).mul(100).round(10).astype(str) + '%')

print("Value counts for Gender:")
print(df2["PatientGender_new"].value_counts())
print("\n")
print("Value counts for Race:")
print(df2["RaceCodeList_new"].value_counts())
print("\n")
print("Value counts for Ethnicity:")
print(df2["patientEthnicity_new"].value_counts())

## Now merging with the Wound_assessment table with the imputed information

new_df_2=df2[["PatientID","PatientGender_new","RaceCodeList_new","patientEthnicity_new"]]
new_df=pd.merge(Wound_assessment,new_df_2,on="PatientID")
#new_df
new_df

new_df["PatientGender_new"]=new_df["PatientGender_new"].replace({0.0:"M",1.0:"F"})
new_df["RaceCodeList_new"]=new_df["RaceCodeList_new"].replace({0.0:"White",1.0:"Black",2.0:"Asian",3.0:"ASKU",
                                                 4.0:"American Indian or Alaska Native",
                                                 5.0:"Native Hawaiian or Other Pacific Islander",
                                                 6.0:"Other"})
new_df["patientEthnicity_new"]=new_df["patientEthnicity_new"].replace({0.0:"Non-Hispanic",1.0:"Hispanic",2.0:"Patient Declined",3.0:"ASKU"})
new_df=new_df.drop(["PatientGender","RaceCodeList","patientEthnicity"],axis=1)
new_df.isnull().sum()

Wound_assessment=new_df
list_patients=Wound_assessment["PatientID"].unique()

unique_wound_number=0
for i in range(len(list_patients)):
    #print(i)
#     if i>9:
#         break
    patient=list_patients[i]
    df_patient=Wound_assessment.query("PatientID == @patient")
    if df_patient.shape[0]>1:
        #print(i)
        wounds_of_list_patient=df_patient["WoundNumber"].unique().tolist()
        #print(wounds_of_list_patient)
        indivi_wound=0
        for j in range(len(wounds_of_list_patient)):
            wound_number= wounds_of_list_patient[j]
            df_patient_1_wound=df_patient.query("WoundNumber==@wound_number")
            #display(df_patient_1_wound)
            if df_patient_1_wound.shape[0]>1:
                if df_patient_1_wound.dropna().shape[0]>=1:
                    indivi_wound=indivi_wound+1
        print("For patient",i+1,"number of unique wounds for which forecasting can be done:",indivi_wound)
#                     #

#                     w_length=pd.Series(df_patient_1_wound["Length"].values)
#                     y = w_length.dropna().values
#                     x=w_length.dropna().index.values
#                     cs = CS(x,y)
#                     x1=w_length.index.values
#                     y1=cs(w_length.index.values)
#                     df_patient_1_wound["Length"]=y1


#                     w_width=pd.Series(df_patient_1_wound["Width"].values)
#                     y = w_width.dropna().values
#                     x=w_width.dropna().index.values
#                     cs = CS(x,y)
#                     x1=w_width.index.values
#                     y1=cs(w_width.index.values)
#                     df_patient_1_wound["Width"]=y1
                    
#                     w_depth=pd.Series(df_patient_1_wound["Depth"].values)
#                     y = w_depth.dropna().values
#                     x=w_depth.dropna().index.values
#                     cs = CS(x,y)
#                     x1=w_depth.index.values
#                     y1=cs(w_depth.index.values)
#                     df_patient_1_wound["Depth"]=y1
                    
                    
#                     display(df_patient_1_wound)
        #print("For patient",i,"number of unique wounds",len(wound_list_patient))
        #for j in range(df_patient_specific_wound=
        unique_wound_number=unique_wound_number+indivi_wound
        #list_count_patient_for_forecasting.append(i)
        #display(df_patient)
print("Total unique wounds:",unique_wound_number)

new_df

new_df.to_csv("3_imputed_data_for_cleaning.csv")

## Checking if there is statistically significant difference in the race distrbution before and after imputation

race_dis=pd.read_csv("race_dis.csv")
race_dis

x=race_dis["Count before impute"].values
x

y=race_dis["Count after impute"].values
y

stats.kruskal(x, y)
